# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateProjectRequest:
    properties:
      branch:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: string
      internal_port:
        type: integer
      internalPort:
        type: integer
      name:
        type: string
      repository_url:
        type: string
      repositoryUrl:
        type: string
      teamId:
        type: string
    type: object
  models.CreateTokenRequest:
    properties:
      environmentId:
        type: string
      expiresAt:
        type: string
      ipAllowlist:
        items:
          type: string
        type: array
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
...
  models.ProjectMember:
    properties:
      acceptedAt:
        type: string
      email:
        type: string
      id:
        type: string
      invitedAt:
        type: string
      permission:
        type: string
      projectId:
        type: string
      status:
        type: string
      userId:
        type: string
    type: object
...
  models.Webhook:
    properties:
      createdAt:
        type: string
      eventTypes:
        items:
          type: string
        type: array
      id:
        type: string
      includePrEnvironments:
        type: boolean
      projectId:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Vessl API Documentation
  title: Vessl API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      consumes:
      - application/json
      description: ListProjects endpoint
      produces:
      - application/json
      responses: {}
      summary: ListProjects endpoint
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: CreateProject endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRequest'
      produces:
      - application/json
      responses: {}
      summary: CreateProject endpoint
      tags:
      - Projects
  /api/projects/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteProject endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeleteProject endpoint
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: GetProject endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetProject endpoint
      tags:
      - Projects
...
  /api/projects/{projectId}/members:
    get:
      consumes:
      - application/json
      description: ListMembers endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListMembers endpoint
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: AddMember endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProjectMember'
      produces:
      - application/json
      responses: {}
      summary: AddMember endpoint
      tags:
      - Projects
  /api/projects/{projectId}/members/{id}:
    delete:
      consumes:
      - application/json
      - application/json
      description: |-
        RemoveMember endpoint
        Remove Project Member
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses: {}
      summary: Remove Project Member
      tags:
      - Teams
      - Projects
  /api/projects/{projectId}/tokens:
    get:
      consumes:
      - application/json
      description: ListTokens endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListTokens endpoint
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: CreateToken endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTokenRequest'
      produces:
      - application/json
      responses: {}
      summary: CreateToken endpoint
      tags:
      - Projects
  /api/projects/{projectId}/tokens/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteToken endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeleteToken endpoint
      tags:
      - Projects
  /api/projects/{projectId}/webhooks:
    get:
      consumes:
      - application/json
      description: ListWebhooks endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListWebhooks endpoint
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: CreateWebhook endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Webhook'
      produces:
      - application/json
      responses: {}
      summary: CreateWebhook endpoint
      tags:
      - Projects
  /api/projects/{projectId}/webhooks/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteWebhook endpoint
      parameters:
      - description: projectId
        in: path
        name: projectId
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeleteWebhook endpoint
      tags:
      - Projects
...
  /api/services/{serviceId}/serverless/code:
    get:
      consumes:
      - application/json
      description: GetCode endpoint
      parameters:
      - description: serviceId
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetCode endpoint
      tags:
      - AppServices
    post:
      consumes:
      - application/json
      description: SaveCode endpoint
      parameters:
      - description: serviceId
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: SaveCode endpoint
      tags:
      - AppServices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"